moved metrics config api to application module
