+++ moove/api/src/main/kotlin/io/charlescd/moove/api/controller/V2WorkspaceController.kt
@@ -17,6 +17,7 @@
 package io.charlescd.moove.api.controller
 
 import io.charlescd.moove.application.ResourcePageResponse
+import io.charlescd.moove.application.metric.VerifyProviderConnectionInteractor
 import io.charlescd.moove.application.metric.VerifyWorkspaceProviderConnectionInteractor
 import io.charlescd.moove.application.metric.response.ProviderConnectionResponse
 import io.charlescd.moove.application.user.response.UserResponse
@@ -27,6 +28,7 @@ import io.charlescd.moove.application.workspace.request.PatchWorkspaceRequest
 import io.charlescd.moove.application.workspace.response.WorkspaceResponse
 import io.charlescd.moove.domain.MetricConfiguration
 import io.charlescd.moove.domain.PageRequest
+import io.charlescd.moove.metrics.api.response.ProviderConnectionRepresentation
 import io.swagger.annotations.ApiImplicitParam
 import io.swagger.annotations.ApiOperation
 import javax.validation.Valid
@@ -43,6 +45,7 @@ class V2WorkspaceController(
     private val findWorkspaceInteractor: FindWorkspaceInteractor,
     private val findAllWorkspacesInteractor: FindAllWorkspaceInteractor,
     private val findAllWorkspaceUsersInteractor: FindAllWorkspaceUsersInteractor,
+    private val verifyProviderConnectionInteractor: VerifyProviderConnectionInteractor,
     private val verifyWorkspaceProviderConnectionInteractor: VerifyWorkspaceProviderConnectionInteractor
 ) {
 
@@ -130,6 +133,15 @@ class V2WorkspaceController(
         return this.findAllWorkspaceUsersInteractor.execute(workspaceId, name, email, pageable)
     }
 
+    @ApiOperation(value = "Verify metrics provider connection")
+    @GetMapping("/config/verify-provider-connection")
+    fun verifyProviderConnection(
+        @RequestParam provider: String,
+        @RequestParam providerType: MetricConfiguration.ProviderEnum
+    ): ProviderConnectionResponse =
+        verifyProviderConnectionInteractor.execute(provider, providerType)
+
+    @ApiOperation(value = "Verify metrics provider connection to the given Workspace")
     @GetMapping("/{workspaceId}/config/verify-provider-connection")
     fun verifyProviderConnectionByIdAndWorkspaceId(
         @PathVariable workspaceId: String,
